requirements.txt
autopep8==1.4.4
dj-database-url==0.5.0
Django==2.2.5
django-cors-headers==3.1.0
django-heroku==0.3.1
djangorestframework==3.10.3
gunicorn==19.9.0
psycopg2==2.8.3
pycodestyle==2.5.0
pytz==2019.2
sqlparse==0.3.0
whitenoise==4.1.4
